requirements.txt
boto==2.43.0
dj-database-url==0.4.1
Django==1.8.5
django-embed-video==1.1.0
django-storages==1.5.1
django-taggit==0.17.1
Markdown==2.6.2
Pillow==2.9.0
psycopg2==2.6.2

pytz==2016.7
requests==2.11.1
